backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
qdrant-client==1.9.1
sentence-transformers==3.0.1
pypdf==4.2.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
gunicorn==22.0.0
uvicorn==0.30.0
Pillow==10.4.0
pytesseract==0.3.13
